refactor: modified unnecessafy field and name
