+++ model/src/main/java/com/apighost/model/loadtest/parameter/LoadTestExecuteParameter.java
@@ -1,6 +1,7 @@
 package com.apighost.model.loadtest.parameter;
 
 import com.apighost.model.scenario.Scenario;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import java.util.List;
@@ -18,13 +19,16 @@ public class LoadTestExecuteParameter {
 
     private final String name;
     private final String description;
-    private final LoadTest loadTest;
+    private final long thinkTimeMs;
+    @JsonProperty("stage")
+    private final List<Stage> stages;
     private final List<Scenario> scenarios;
 
     private LoadTestExecuteParameter(LoadTestExecuteParameter.Builder builder) {
         this.name = builder.name;
         this.description = builder.description;
-        this.loadTest = builder.loadTest;
+        this.thinkTimeMs = builder.thinkTimeMs;
+        this.stages = builder.stages;
         this.scenarios = builder.scenarios;
     }
 
@@ -36,7 +40,8 @@ public class LoadTestExecuteParameter {
 
         private String name;
         private String description;
-        private LoadTest loadTest;
+        private long thinkTimeMs;
+        private List<Stage> stages;
         private List<Scenario> scenarios;
 
         public LoadTestExecuteParameter.Builder name(String name) {
@@ -49,8 +54,14 @@ public class LoadTestExecuteParameter {
             return this;
         }
 
-        public LoadTestExecuteParameter.Builder loadTest(LoadTest loadTest) {
-            this.loadTest = loadTest;
+        public LoadTestExecuteParameter.Builder thinkTimeMs(long thinkTimeMs) {
+            this.thinkTimeMs = thinkTimeMs;
+            return this;
+        }
+
+        @JsonProperty("stage")
+        public LoadTestExecuteParameter.Builder stages(List<Stage> stages) {
+            this.stages = stages;
             return this;
         }
 
@@ -75,8 +86,13 @@ public class LoadTestExecuteParameter {
         return description;
     }
 
-    public LoadTest getLoadTest() {
-        return loadTest;
+    public long getThinkTimeMs() {
+        return thinkTimeMs;
+    }
+
+    @JsonProperty("stage")
+    public List<Stage> getStages() {
+        return stages;
     }
 
     public List<Scenario> getScenarios() {
