+++ model/src/main/java/com/apighost/model/loadtest/parameter/LoadTestParameter.java
@@ -1,5 +1,6 @@
 package com.apighost.model.loadtest.parameter;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import java.util.List;
@@ -16,13 +17,16 @@ public class LoadTestParameter {
 
     private final String name;
     private final String description;
-    private final LoadTest loadTest;
+    private final long thinkTimeMs;
+    @JsonProperty("stage")
+    private final List<Stage> stages;
     private final List<String> scenarios;
 
     private LoadTestParameter(Builder builder) {
         this.name = builder.name;
         this.description = builder.description;
-        this.loadTest = builder.loadTest;
+        this.thinkTimeMs = builder.thinkTimeMs;
+        this.stages = builder.stages;
         this.scenarios = builder.scenarios;
     }
 
@@ -34,7 +38,8 @@ public class LoadTestParameter {
 
         private String name;
         private String description;
-        private LoadTest loadTest;
+        private long thinkTimeMs;
+        private List<Stage> stages;
         private List<String> scenarios;
 
         public Builder name(String name) {
@@ -47,8 +52,14 @@ public class LoadTestParameter {
             return this;
         }
 
-        public Builder loadTest(LoadTest loadTest) {
-            this.loadTest = loadTest;
+        public Builder thinkTimeMs(long thinkTimeMs) {
+            this.thinkTimeMs = thinkTimeMs;
+            return this;
+        }
+
+        @JsonProperty("stage")
+        public Builder stages(List<Stage> stages) {
+            this.stages = stages;
             return this;
         }
 
@@ -73,8 +84,13 @@ public class LoadTestParameter {
         return description;
     }
 
-    public LoadTest getLoadTest() {
-        return loadTest;
+    public long getThinkTimeMs() {
+        return thinkTimeMs;
+    }
+
+    @JsonProperty("stage")
+    public List<Stage> getStages() {
+        return stages;
     }
 
     public List<String> getScenarios() {
